Use systemd-sysusers to create user & group
This duplicates the sysusers config from upstream but only installs the
upstream one. The configs should be kept in sync.

We can not use the upstream one as it requires pre-processing that will
only happen at a later time in the build process.

See: https://docs.fedoraproject.org/en-US/packaging-guidelines/UsersAndGroups/#_dynamic_allocation

https://src.fedoraproject.org/rpms/flatpak/pull-request/8